VISITOR 1. Link Read More kurang nama controller OK 2. Gambar Upload kesalahan nyimpan name gambar, form edit kurang enctype="multipart/form-data" 3. Gambar View OK 4. View Post Gambar OK 5. Insert Comment OK 6. List Comment OK 7. Hapus Comment Oleh admin OK 8. Rapihkan Tampilan a. Home ok b. List Post ok c. View Post ok d. List comment ok 9. Pagination Ok ADMIN 1. Preview Post OK 2. Hapus Post OK 3. List Comment OK 4. View Comment OK 5. Hapus Comment OK

+++ app/Http/Controllers/PostController.php
@@ -42,12 +42,13 @@ class PostController extends Controller
         $post->title = $request->get('title');
         $post->desc = $request->get('content');
        if($request->hasFile('image')){
-         $post->image = $request->file('image');
+         // $post->image = $request->file('image');
 
          $imageName = time().'.'.$request->image->getClientOriginalExtension();
+         $post->image = $imageName;
          $request->image->move(public_path('/images'), $imageName);
         }else{
-         $post->image = ('images/default.jpg');
+         $post->image = ('default.jpg');
         }
         $post->save();
         return redirect('admin/post');
@@ -61,7 +62,9 @@ class PostController extends Controller
      */
     public function show($id)
     {
-        //
+        $post = post::find($id);
+        
+        return view('admin/posting-preview')->with('data',$post);
     }
 
     /**
@@ -89,9 +92,9 @@ class PostController extends Controller
        $post->title = $request->get('title');
        $post->desc = $request->get('content');
        if($request->hasFile('image')){
-         $post->image = $request->file('image');
-         
+
          $imageName = time().'.'.$request->image->getClientOriginalExtension();
+         $post->image = $imageName;
          $request->image->move(public_path('/images'), $imageName);
         }
        $post->save();
@@ -109,12 +112,15 @@ class PostController extends Controller
 
     public function destroyall(Request $request)
     {
-        if(count(collect($request->checkbox)) > 1){
-          $post = Post::whereIn('id',$request->get('checkbox'));
-          $post->delete();
-        }else{
-          $post = post::find($request->get('checkbox'))->first();
-          $post->delete();
+        if($request->get('checkbox') !== null)
+        {
+            if(count(collect($request->checkbox)) > 1){
+              $post = Post::whereIn('id',$request->get('checkbox'));
+              $post->delete();
+          }else{
+              $post = post::find($request->get('checkbox'))->first();
+              $post->delete();
+          }
         }
         return back();
     }
